requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
google-search-results
beautifulsoup4
pyyaml
python-dotenv
langchain-experimental>=0.0.60
 
# Neo4j GraphRAG
neo4j>=5.14.0
langchain-neo4j>=0.1.0
 
# PDF processing
pypdf>=4.0.0
pdfplumber>=0.11.0
 
# Embeddings & LLM
openai>=1.30.0
tiktoken>=0.7.0
 
# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0